backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
redis[hiredis]==5.2.1
httpx==0.28.1
python-dotenv==1.0.1
apscheduler==3.10.4
playwright==1.49.1
beautifulsoup4==4.12.3
lxml==5.3.0
structlog==24.4.0
tenacity==9.0.0
email-validator==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==8.3.4
pytest-asyncio==0.25.0
